Move expansion and file utils tests
