+++ crates/triplets-hf-source/src/file_utils.rs
@@ -48,6 +48,20 @@ mod tests {
     use std::path::Path;
 
     #[test]
+    fn is_gzip_path_detects_gz_extension() {
+        assert!(is_gzip_path(Path::new("file.jsonl.gz")));
+        assert!(is_gzip_path(Path::new("file.GZ")));
+        assert!(is_gzip_path(Path::new("file.Gz")));
+        assert!(is_gzip_path(Path::new("file.tar.gz")));
+        assert!(!is_gzip_path(Path::new("file.parquet")));
+        assert!(!is_gzip_path(Path::new("file.jsonl")));
+        assert!(!is_gzip_path(Path::new("file.simdr")));
+        assert!(!is_gzip_path(Path::new("no-extension")));
+        assert!(!is_gzip_path(Path::new("file.gz.bak")));
+        assert!(!is_gzip_path(Path::new("")));
+    }
+
+    #[test]
     fn resolve_inner_extension_handles_compound_extensions() {
         // Compound extensions
         assert_eq!(
